header styles and position
