--- src/components/Header.jsx
@@ -1,26 +1,0 @@
-import { Fragment } from 'react';
-
-import logo from '../img/logo.png';
-import classes from './Header.module.css';
-
-function Header() {
-  return (
-    <Fragment>
-      {/* <header>
-        <img className={classes.logo} src={logo} alt="Logo" />
-        <h2>Algorithm Visualizer</h2>
-        <nav>
-          <button>Sorting Algorithms</button>
-          <ul>
-            <li>Algorithms</li>
-            <li>Algorithms</li>
-            <li>Algorithms</li>
-          </ul>
-          <button>Visualize</button>
-        </nav>
-      </header> */}
-    </Fragment>
-  );
-}
-
-export default Header;
