--- src/components/Header.module.css
@@ -1,8 +1,0 @@
-.logo {
-  height: 50px;
-  width: 50px;
-}
-
-h2 {
-  opacity: 86%;
-}
