+++ src/components/header/Header.module.css
@@ -1,0 +1,18 @@
+h2 {
+  opacity: 86%;
+  font-size: 2em;
+  margin-top: 1rem;
+}
+
+.header {
+  background-color: var(--secondary-background);
+  display: flex;
+  position: relative;
+  padding: 0.75rem;
+}
+
+.logo {
+  height: 5em;
+  width: 5em;
+  margin: 0 2em;
+}
