+++ src/components/header/NavBar.module.css
@@ -1,0 +1,82 @@
+button {
+  border-radius: 1rem;
+  background: none;
+  font-size: 2em;
+  padding: 1rem;
+  border: none;
+  width: 10em;
+  color: var(--main-font-color);
+  cursor: pointer;
+}
+
+button:focus {
+  outline: none;
+  background: var(--hover-color);
+}
+
+.navlist {
+  display: flex;
+  align-items: flex-start;
+  position: absolute;
+  margin-top: 0.5em;
+  left: 20%;
+}
+
+.btn--primary {
+  background: #d3d4d8;
+  color: var(--dark-background);
+  margin-left: 5em;
+}
+
+.btn--primary:focus,
+.btn--primary:hover {
+  background: #e4e6eb;
+}
+
+.btn--selected {
+  background: var(--hover-color);
+}
+
+.dropdown {
+  width: 33%;
+  list-style: none;
+  border-radius: 1rem;
+  position: relative;
+  display: flex;
+  margin-left: 1em;
+  align-items: flex-start;
+  flex-direction: column;
+  height: auto;
+}
+
+.dropdown:hover {
+  background: var(--hover-color);
+  cursor: pointer;
+}
+
+.dropdown--btn {
+  margin: 0.3em 0.5em;
+  font-size: 2em;
+  cursor: pointer;
+}
+
+.arrow {
+  font-size: 0.7em;
+  margin-left: 1em;
+}
+
+.dropdown--item {
+  font-size: 2em;
+  padding: 0.5em;
+  border-bottom-style: solid;
+  border-width: 1px;
+  width: 100%;
+}
+
+.dropdown--item:hover {
+  background-color: var(--secondary-background);
+}
+
+.dropdown--item.last--item {
+  border-bottom: none;
+}
