+++ src/index.css
@@ -2,7 +2,10 @@
 
 :root {
   --dark-background: #121212;
-  --main-font-color: #ffffff;
+  --secondary-background: #242526;
+  --main-font-color: #e4e6eb;
+  --hover-color: #3a3b3c;
+  --secondary-color: #b0b3b8;
   --font-family: 'Open Sans', 'Roboto', 'Segoe UI', 'Oxygen', sans-serif;
 }
 
@@ -10,6 +13,10 @@
   margin: 0;
   padding: 0;
   box-sizing: border-box;
+}
+
+html {
+  font-size: 62.5%;
 }
 
 body {
